requirements.txt
mysql-connector-python
google-api-python-client
google-auth
python-dotenv
